Implement persistent storage
Just writing and reading to a json file.
Types could definitely be cleaned up.
Had to do some refactoring to allow stored data to be serializable.
Need to be careful about message configs that need to trigger side effects
when set. It's not currently an issue but the way things are written right
now it could be.